Workarounded timer delay caused by async_read.
Since it's not possible to interrupt read in event loop, we have to
create separate thread to run heartbeat and keepalive timers. However
we don't explicitly run timers in it, instead we run ASIO event loop in
it, but due to polling "strict ordering" hack it can't block both
threads.

To hide this detail we added method Client::run, that creates second
thread with event loop. However documentation describes what really
happens in it. So we allow user to not use Client::run and create
threads explicitly (and more than 2, if user needs it).